@@ -10,9 +10,9 @@
 
             <div class="py-3">
 
-                <x-button.basic text="Nuevo proyecto"
+                <x-button.basic text="New project"
                                 icon="plus-circle"
-                                :url="action([\App\Http\Controllers\ProjectController::class,'form'])"  />
+                                :url="action([\App\Http\Controllers\ProjectController::class,'new'])"  />
             </div>
 
 
@@ -33,10 +33,7 @@
                                         <th class="text-left py-3 px-4 font-semibold text-sm">Fecha</th>
                                         <th class="text-left py-3 px-4 font-semibold text-sm">
 
-                                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 6v6m0 0v6m0-6h6m-6 0H6" />
-                                            </svg>
-                                            <span>&nbsp;</span>
+                                            <x-icons.plus-circle/>
 
                                         </th>
                                     </tr>
@@ -48,7 +45,7 @@
                                             <td class="w-1/3 text-left py-3 px-4">{{$project->client->name}}</td>
                                             <td class="text-left py-3 px-4">{{$project->date}}</td>
                                             <td>
-                                                <span>a bottom</span>
+                                                <x-button.basic :url="action('ProjectController@edit',$project->id)" icon="pencil" text="Edit" />
                                             </td>
                                         </tr>
                                         @endforeach
